rename: 재사용을 위한 폴더명 수정(profile)
